requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.0 